Re-implemented color handling to mimic Processing.
Original idea: define Processing color type as a libgdx Color:
class color extends Color {}

An advantage to this approach is that is allows use of Java's type
system to distinguish between ARGB and greyscale.

Disadvantage: extra layer of indirection, including memory management of
the Color object.

Colors are now represented by ARGB8888 ints in PApplet client
subclasses, just as they are in Processing.  Color functions now follow
the same logic as processing to determine whether an int should be
interpeted as ARGB or greyscale.